Some changes in faculty and student service

+++ backend/src/main/java/com/netcracker/devschool/dev4/etalon/service/FacultyServiceImpl.java
@@ -1,16 +1,25 @@
 package com.netcracker.devschool.dev4.etalon.service;
 
 import com.netcracker.devschool.dev4.etalon.entity.Faculty;
+import com.netcracker.devschool.dev4.etalon.entity.Student;
 import com.netcracker.devschool.dev4.etalon.repository.FacultyRepository;
+import com.netcracker.devschool.dev4.etalon.repository.SpecialityRepository;
 import org.springframework.stereotype.Service;
 
 import javax.annotation.Resource;
+import javax.swing.*;
 import java.util.List;
 
 @Service
 public class FacultyServiceImpl  implements FacultyService {
     @Resource
     private FacultyRepository facultyRepository;
+
+    @Resource
+    private SpecialityRepository specialityRepository;
+
+    @Resource
+    private StudentService studentService;
 
     @Override
     public Faculty create(Faculty faculty) {
@@ -19,7 +28,13 @@ public class FacultyServiceImpl  implements FacultyService {
 
     @Override
     public void deleteFacultyById(int id){
-        facultyRepository.delete(id);
+        List<Student> students = studentService.findStudentByIdFaculty(id);
+        if (students.isEmpty()) {
+            specialityRepository.deleteSpecialityByFaculty(id);
+            facultyRepository.delete(id);
+        } else {
+            JOptionPane.showMessageDialog(null, "Переведите всех студентов на другой факультет!");
+        }
     }
 
     @Override
